pyproject.toml
[tool.poetry]
name = "calmzeus"
version = "0.0.1"
description = "Exploration for project 'peacefulares'"
authors = ["lo-b <bram_lo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
notebook = ">=6.5.1,<7.0.0"
faiss-cpu = "^1.8.0.post1"
langchain = "^0.3.2"
voyageai = "^0.2.3"
load-dotenv = "^0.1.0"
langchain-community = "^0.3.1"
tree-sitter = "0.21.3"
tree-sitter-languages = "^1.10.2"
langchain-voyageai = "^0.1.2"
rich = "^13.9.2"
langgraph = "^0.2.35"
langsmith = "^0.1.133"
langchain-ollama = "^0.2.0"
langchain-qdrant = "^0.1.4"


[tool.poetry.group.dev.dependencies]
jupyter-ascending = "^0.1.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
